+++ test/bigtest/tests/settings/setting-suffixes-test.js
@@ -12,8 +12,9 @@ describe('Setting of Closing Reasons', function () {
   const suffixesSetting = new SuffixesInteractor();
   const deleteConfirmation = new ConfirmationInteractor('#delete-controlled-vocab-entry-confirmation');
 
-  beforeEach(function () {
+  beforeEach(async function () {
     this.visit('/settings/orders/suffixes');
+    await suffixesSetting.whenLoaded();
   });
 
   it('should be rendered', () => {
@@ -24,6 +25,7 @@ describe('Setting of Closing Reasons', function () {
   describe('Add new suffix', function () {
     beforeEach(async function () {
       await suffixesSetting.addSuffixBtn.click();
+      await suffixesSetting.whenLoaded();
     });
 
     it('renders fields for new suffix', () => {
@@ -34,6 +36,7 @@ describe('Setting of Closing Reasons', function () {
     describe('Cancel add new suffix', function () {
       beforeEach(async function () {
         await suffixesSetting.suffixes.list(0).cancelButton.click();
+        await suffixesSetting.whenLoaded();
       });
 
       it('renders fields for new suffix', () => {
@@ -45,6 +48,7 @@ describe('Setting of Closing Reasons', function () {
       beforeEach(async function () {
         await suffixesSetting.suffixes.list(0).nameInput.fill('test');
         await suffixesSetting.suffixes.list(0).saveButton.click();
+        await suffixesSetting.whenLoaded();
       });
 
       it('renders saved suffix', () => {
@@ -56,6 +60,7 @@ describe('Setting of Closing Reasons', function () {
           await suffixesSetting.suffixes.list(0).editButton.click();
           await suffixesSetting.suffixes.list(0).nameInput.fill('test new');
           await suffixesSetting.suffixes.list(0).saveButton.click();
+          await suffixesSetting.whenLoaded();
         });
 
         it('renders edited suffix', () => {
@@ -67,6 +72,7 @@ describe('Setting of Closing Reasons', function () {
         beforeEach(async function () {
           await suffixesSetting.suffixes.list(0).deleteButton.click();
           await deleteConfirmation.confirm();
+          await suffixesSetting.whenLoaded();
         });
 
         it('does not render empty list', () => {